Add options for color lookup.
