+++ src/ImageSharp/Processing/Processors/Quantization/QuantizerOptions.cs
@@ -38,4 +38,9 @@ public class QuantizerOptions
         get => this.maxColors;
         set => this.maxColors = Numerics.Clamp(value, QuantizerConstants.MinColors, QuantizerConstants.MaxColors);
     }
+
+    /// <summary>
+    /// Gets or sets the color matching mode used for matching pixel values to palette colors.
+    /// </summary>
+    public ColorMatchingMode ColorMatchingMode { get; set; } = ColorMatchingMode.Hybrid;
 }
