--- src/ImageSharp/Properties/AssemblyInfo.cs
@@ -1,9 +1,0 @@
-﻿// Copyright (c) Six Labors.
-// Licensed under the Six Labors Split License.
-
-// Redundant suppressing of SA1413 for Rider.
-[assembly:
-    System.Diagnostics.CodeAnalysis.SuppressMessage(
-        "StyleCop.CSharp.MaintainabilityRules",
-        "SA1413:UseTrailingCommasInMultiLineInitializers",
-        Justification = "Follows SixLabors.ruleset")]
